issue-1038: implicit class StructSDLB
