+++ sdl-core/src/test/scala/io/smartdatalake/util/spark/dataset/EqualityTest.scala
@@ -20,9 +20,9 @@
 package io.smartdatalake.util.spark.dataset
 
 import io.smartdatalake.testutils.TestTool
-import io.smartdatalake.util.spark.GetSession.{createSparkSession, loggEnv}
 import io.smartdatalake.testutils.spark.dataset.Collection._
 import io.smartdatalake.util.PrecisionDef
+import io.smartdatalake.util.spark.GetSession.{createSparkSession, loggEnv}
 import org.apache.spark.sql.functions.col
 import org.apache.spark.sql.types._
 import org.apache.spark.sql.{DataFrame, Row, SparkSession}
@@ -42,71 +42,6 @@ class EqualityTest extends AnyFlatSpec with Matchers with ScalaCheckPropertyChec
   import spark.implicits._
 
   loggEnv
-
-  "schemataEqual equals" should "fail if column count differ" in {
-    val schL: StructType = StructType(Array(
-      StructField(name = "x", dataType = IntegerType, nullable = true),
-      StructField(name = "y", dataType = IntegerType, nullable = true)))
-    val schR: StructType = StructType(Array(StructField(name = "x", dataType = IntegerType, nullable = true)))
-    schemataEqual(schL, schR) shouldBe false
-  }
-
-  "Schemata equal" should "pass even if column order differ" in {
-    val schL: StructType = StructType(Array(
-      StructField(name = "x", dataType = IntegerType, nullable = true),
-      StructField(name = "y", dataType = IntegerType, nullable = true)))
-    val schR: StructType = StructType(Array(
-      StructField(name = "y", dataType = IntegerType, nullable = true),
-      StructField(name = "x", dataType = IntegerType, nullable = true)))
-    schemataEqual(schL, schR) shouldBe true
-  }
-
-  "Schemata equal" should "pass even if nullability differs" in {
-    val schL: StructType = StructType(Array(
-      StructField(name = "x", dataType = IntegerType, nullable = true),
-      StructField(name = "y", dataType = IntegerType, nullable = true)))
-    val schR: StructType = StructType(Array(
-      StructField(name = "x", dataType = IntegerType, nullable = true),
-      StructField(name = "y", dataType = IntegerType, nullable = false)))
-    schemataEqual(schL, schR) shouldBe true
-  }
-
-  "Schemata equal" should "fail if types differ" in {
-    val schL: StructType = StructType(Array(
-      StructField(name = "x", dataType = IntegerType, nullable = true),
-      StructField(name = "y", dataType = DoubleType, nullable = true)))
-    val schR: StructType = StructType(Array(
-      StructField(name = "x", dataType = IntegerType, nullable = true),
-      StructField(name = "y", dataType = FloatType, nullable = true)))
-    schemataEqual(schL, schR) shouldBe false
-  }
-
-  "Schemata equal" should "fail if nested types differ" in {
-    val valuesL: StructType = StructType(Array(
-      StructField(name = "x", dataType = IntegerType, nullable = true),
-      StructField(name = "y", dataType = DoubleType, nullable = true)))
-    val valuesR: StructType = StructType(Array(
-      StructField(name = "x", dataType = IntegerType, nullable = true),
-      StructField(name = "y", dataType = FloatType, nullable = true)))
-    val schL: StructType = StructType(Array(
-      StructField(name = "id", dataType = StringType, nullable = false),
-      StructField(name = "values", dataType = valuesL, nullable = true)))
-    val schR: StructType = StructType(Array(
-      StructField(name = "id", dataType = StringType, nullable = false),
-      StructField(name = "values", dataType = valuesR, nullable = true)))
-    schemataEqual(schL, schR) shouldBe false
-  }
-
-  "Schemata equal" should "ignore nullability but not containsNull" in {
-    val rightSchema = StructType(Array(
-      StructField("code_d", ArrayType(StringType, containsNull = true), nullable = true)
-    ))
-    val leftSchema =
-      StructType(Array(
-        StructField("code_d", ArrayType(StringType, containsNull = false), nullable = true)
-      ))
-    schemataEqual(leftSchema, rightSchema) shouldBe false
-  }
 
   "getSymmetricDifference" should "return empty df if generated df is compared with itself" in {
     forAll(genA = genExactFrame) { df =>
